requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.2
Flask-Login==0.6.2
Werkzeug==2.3.6
requests==2.31.0
tweepy==4.14.0
google-cloud-aiplatform==1.42.0
newsapi-python==0.2.7
schedule==1.2.0
newsapi-python==0.2.7
requests==2.31.0